Ajout de la session

+++ models/Client.php
@@ -9,22 +9,26 @@ class Client extends User {
     private $siret;
     private $adress;
     private $role;
+    private $created;
 
     /**
+     * Constructeur
      * @param string $firstname
      * @param string $lastname
-     * @param mixed string
+     * @param string $email
      * @param string $phone
      * @param string $siret
-     * @param string $adress
+     * @param int $created
+     * @param string|null $adress
      * @param int $role
      */
-    public function __construct (string $firstname, string $lastname, string $email, string $phone, string $siret, string $adress = null, int $role = 3) {
+    public function __construct (string $firstname, string $lastname, string $email, string $phone, string $siret, int $created, string $adress = null, int $role = 3) {
         parent::__construct($firstname, $lastname, $email);
         $this->phone = $phone;
         $this->siret = $siret;
         $this->adress = $adress;
         $this->role = $role;
+        $this->created = $created;
     }
 
     /**
@@ -59,6 +63,14 @@ class Client extends User {
         return $this->role;
     }
 
+    /**
+     * Récupération de l'admin créateur
+     * @return int
+     */
+    public function getCreated () :int {
+        return $this->created;
+    }
+
     /************************************** **************************************/
     /*********************************** CREATE **********************************/
     /************************************** **************************************/
@@ -69,8 +81,9 @@ class Client extends User {
      */
     public function isExist () :bool {
         $databaseConnection = Database::getConnection();
-        $query = $databaseConnection->prepare('SELECT `siret` FROM users WHERE siret = :siret ;');
+        $query = $databaseConnection->prepare('SELECT `siret` FROM `users` WHERE `siret` = :siret AND `Id_role` = 3 AND `created_by` = :created ;');
         $query->bindValue(':siret', $this->getSiret(), PDO::PARAM_STR);
+        $query->bindValue(':created', $this->getCreated(), PDO::PARAM_INT);
         $query->execute();
         $result = $query->fetch(PDO::FETCH_OBJ);
         return $result ? true : false;
@@ -82,15 +95,17 @@ class Client extends User {
      */
     public function add () :bool {
         $databaseConnection = Database::getConnection();
-        $query = $databaseConnection->prepare('INSERT INTO users (firstname, lastname, email, phone, siret, adress, Id_role) VALUES (:firstname, :lastname, :email, :phone, :siret, :adress, :role) ;');
-        $query->bindValue(':firstname', $this->firstname, PDO::PARAM_STR);
-        $query->bindValue(':lastname', $this->lastname, PDO::PARAM_STR);
-        $query->bindValue(':email', $this->email, PDO::PARAM_STR);
-        $query->bindValue(':phone', $this->phone, PDO::PARAM_STR);
-        $query->bindValue(':siret', $this->siret, PDO::PARAM_STR);
-        $query->bindValue(':adress', $this->adress, PDO::PARAM_STR);
-        $query->bindValue(':role', $this->role, PDO::PARAM_INT);
+        $query = $databaseConnection->prepare('INSERT INTO `users` (firstname, lastname, email, phone, siret, adress, Id_role, created_by) VALUES (:firstname, :lastname, :email, :phone, :siret, :adress, :role, :created) ;');
+        $query->bindValue(':firstname', $this->getFirstname(), PDO::PARAM_STR);
+        $query->bindValue(':lastname', $this->getLastname(), PDO::PARAM_STR);
+        $query->bindValue(':email', $this->getEmail(), PDO::PARAM_STR);
+        $query->bindValue(':phone', $this->getPhone(), PDO::PARAM_STR);
+        $query->bindValue(':siret', $this->getSiret(), PDO::PARAM_STR);
+        $query->bindValue(':adress', $this->getAdress(), PDO::PARAM_STR);
+        $query->bindValue(':role', $this->getRole(), PDO::PARAM_INT);
+        $query->bindValue(':created', $this->getCreated(), PDO::PARAM_INT);
         $query->execute();
+
         return $query->rowCount() > 0 ? true : false;
     }
 
@@ -104,7 +119,8 @@ class Client extends User {
      */
     public static function howManyPages () :int {
         $databaseConnection = Database::getConnection();
-        $query = $databaseConnection->prepare('SELECT COUNT(`Id_users`) as total FROM `users` WHERE Id_role = 3 ;');
+        $query = $databaseConnection->prepare('SELECT COUNT(`Id_users`) as total FROM `users` WHERE `Id_role` = 3 and `created_by` = :created ;');
+        $query->bindValue(':created', $_SESSION['id'], PDO::PARAM_INT);
         $query->execute();
         $result = $query->fetch(PDO::FETCH_OBJ);
         $totalPages = intdiv($result->total, 10);
@@ -135,7 +151,8 @@ class Client extends User {
      */
     public static function getAll () :array {
         $databaseConnection = Database::getConnection();
-        $query = $databaseConnection->prepare('SELECT `firstname`, `lastname`, `id` FROM users where Id_role = 2 ;');
+        $query = $databaseConnection->prepare('SELECT `firstname`, `lastname`, `Id_users` FROM `users` where `Id_role` = 3 and `created_by` = :created ;');
+        $query->bindValue(':created', $_SESSION['id'], PDO::PARAM_INT);
         $query->execute();
         $result = $query->fetchAll(PDO::FETCH_OBJ);
         return $result;
@@ -147,7 +164,8 @@ class Client extends User {
      */
     public static function getTen () :array {
         $databaseConnection = Database::getConnection();
-        $query = $databaseConnection->prepare('SELECT `lastname`, `firstname`, `Id_users` FROM `users` WHERE Id_role = 3 ORDER BY `lastname` ASC LIMIT :numberPerPage OFFSET :offset ');
+        $query = $databaseConnection->prepare('SELECT `lastname`, `firstname`, `Id_users` FROM `users` WHERE `Id_role` = 3 and `created_by` = :created ORDER BY `lastname` ASC LIMIT :numberPerPage OFFSET :offset ;');
+        $query->bindValue(':created', $_SESSION['id'], PDO::PARAM_INT);
         $query->bindValue(':numberPerPage', 10, PDO::PARAM_INT);
         $query->bindValue(':offset', (Client::whichPage() - 1) * 10, PDO::PARAM_INT);
         $query->execute();
@@ -161,8 +179,9 @@ class Client extends User {
      */
     public static function checkId (int $id) : bool {
         $databaseConnection = Database::getConnection();
-        $query = $databaseConnection->prepare('SELECT `Id_users` FROM users WHERE Id_users = :id AND Id_role = 3 ;');
+        $query = $databaseConnection->prepare('SELECT `Id_users` FROM `users` WHERE `Id_users` = :id AND `Id_role` = 3 AND `created_by` = :created ;');
         $query->bindValue(':id', $id, PDO::PARAM_INT);
+        $query->bindValue(':created', $_SESSION['id'], PDO::PARAM_INT);
         $query->execute();
         $result = $query->fetch(PDO::FETCH_OBJ);
         return $result ? true : false;
@@ -176,8 +195,9 @@ class Client extends User {
      */
     public static function getOne (int $id) :array {
         $databaseConnection = Database::getConnection();
-        $query = $databaseConnection->prepare('SELECT `firstname`, `lastname`, `email`, `phone`, `siret`, `adress`, `Id_users` FROM users WHERE Id_users = :id AND Id_role = 3 ;');
+        $query = $databaseConnection->prepare('SELECT `firstname`, `lastname`, `email`, `phone`, `siret`, `adress`, `Id_users` FROM `users` WHERE `Id_users` = :id AND `Id_role` = 3 AND `created_by` = :created ;');
         $query->bindValue(':id', $id, PDO::PARAM_INT);
+        $query->bindValue(':created', $_SESSION['id'], PDO::PARAM_INT);
         $query->execute();
         $result = $query->fetchAll(PDO::FETCH_OBJ);
         return $result;
@@ -193,15 +213,17 @@ class Client extends User {
      */
     public function update (int $id) :bool {
         $databaseConnection = Database::getConnection();
-        $query = $databaseConnection->prepare('UPDATE `users` SET `firstname` = :firstname, `lastname` = :lastname, `email` = :email, `phone` = :phone, `siret` = :siret, `adress` = :adress WHERE `Id_users` = :id ;');
+        $query = $databaseConnection->prepare('UPDATE `users` SET `firstname` = :firstname, `lastname` = :lastname, `email` = :email, `phone` = :phone, `siret` = :siret, `adress` = :adress WHERE `Id_users` = :id AND `Id_role` = 3 AND `created_by` = :created ;');
         $query->bindValue(':firstname', $this->getFirstname(), PDO::PARAM_STR);
         $query->bindValue(':lastname', $this->getLastname(), PDO::PARAM_STR);
         $query->bindValue(':email', $this->getEmail(), PDO::PARAM_STR);
         $query->bindValue(':phone', $this->getPhone(), PDO::PARAM_STR);
-        $query->bindValue(':income', $this->getSiret(), PDO::PARAM_STR);
+        $query->bindValue(':siret', $this->getSiret(), PDO::PARAM_STR);
         $query->bindValue(':adress', $this->getAdress(), PDO::PARAM_STR);
         $query->bindValue(':id', $id, PDO::PARAM_INT);
+        $query->bindValue(':created', $this->getCreated());
         $query->execute();
+        var_dump($query->rowCount());
         return $query->rowCount() > 0 ? true : false;
     }
 
@@ -215,8 +237,9 @@ class Client extends User {
      */
     public static function delete (int $id) :bool {
         $databaseConnection = Database::getConnection();
-        $query = $databaseConnection->prepare('DELETE FROM `users` WHERE `Id_users` = :id ;');
+        $query = $databaseConnection->prepare('DELETE FROM `users` WHERE `Id_users` = :id AND `Id_role` = 3 AND `created_by` = :created ;');
         $query->bindValue(':id', $id, PDO::PARAM_INT);
+        $query->bindValue(':created', $_SESSION['id'], PDO::PARAM_INT);
         $query->execute();
         return $query->rowCount() > 0 ? true : false;
     }
